mp4 to avi lossless audio converter
This is a lossless audio .mp4 to .avi converter that scrapes the CanAm2021 folder, saves the .mp4 files into a local temporary folder, converts them to .avi wtih lossless audio, and then copies it back to the network drive. This script was necessary to do further audio testing on those trimmed files. ffmpeg must be installed and set as an environmental variable for the script to work properly.